rename TracingConfigurationException
