+++ elastic-job-lite-tracing/elastic-job-lite-tracing-api/src/main/java/org/apache/shardingsphere/elasticjob/lite/tracing/JobEventBus.java
@@ -23,6 +23,7 @@ import com.google.common.eventbus.EventBus;
 import com.google.common.util.concurrent.MoreExecutors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.concurrent.BasicThreadFactory;
+import org.apache.shardingsphere.elasticjob.lite.tracing.exception.TracingConfigurationException;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.LinkedBlockingQueue;
@@ -67,7 +68,7 @@ public final class JobEventBus {
         try {
             eventBus.register(jobEventConfig.createJobEventListener());
             isRegistered = true;
-        } catch (final JobEventListenerConfigurationException ex) {
+        } catch (final TracingConfigurationException ex) {
             log.error("Elastic job: create JobEventListener failure, error is: ", ex);
         }
     }
